Moved all cell components into a subfolder in the pricing tab folder
